split index functionality to app.js
